All functions return Message and error.
This is to enable a workflow I need, where an API can receive a message from a client,
add a job into a queue, then return the ID of the job to the client. This is useful for
allowing the client to request updates on the progress of that job.

When implementing this, it was necessary to shared the Message and ID structs, so I
moved them into the "common" dir.

I had not envisaged so many changes when I started, but the change to function
signatures cascaded throughout much of the codebase.

I'm new to Go, so if this is not idiomatic I'm happy to collaborate on further changes.