Don't wait for ticker for first receive in `Queue.ReceiveAndWait` (#69)
Currently ReceiveAndWait first starts a ticker, and then on a tick,
calls q.Receive.

As a side effect, while using the jobs/Runner, there is a mandatory
delay of pollinterval even if there are more jobs that can be
immediately picked up.

This patch, calls q.Receive before waiting on ticker, if nothing is
returned, continues with setting up the ticker and calling q.Receive in
a loop.